feat(plugins): group some EC2 API classes more logically
