--- keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/Action.kt
@@ -1,8 +1,0 @@
-package com.netflix.spinnaker.keel.api.ec2
-
-data class Action(
-  val type: String,
-  val order: Int,
-  val targetGroupName: String?,
-  val redirectConfig: RedirectConfig?
-)
