--- keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/BuildInfo.kt
@@ -1,5 +1,0 @@
-package com.netflix.spinnaker.keel.api.ec2
-
-data class BuildInfo(
-  val packageName: String?
-)
