--- keel-ec2-api/src/main/kotlin/com/netflix/spinnaker/keel/api/ec2/Condition.kt
@@ -1,6 +1,0 @@
-package com.netflix.spinnaker.keel.api.ec2
-
-data class Condition(
-  val field: String,
-  val values: List<String>
-)
